devdash/backend/requirements.txt
# backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
sqlalchemy==2.0.25
alembic==1.13.0
anthropic==0.16.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-jose[cryptography] == 3.4.0
passlib==1.7.4
bcrypt==4.0.1
pytest == 8.4.1
pytest-asyncio == 1.0.0